fix (dev_tests) refactored project structure
